Fix working with S3 blobstore
ArArchiveInputStream in commons-compress 1.18 has a bug that
prevents working with any source stream that does not implement
available() as returning the number of bytes remaining to be
read in the entire stream. Since socket based streams don't
behave like this, it does not work with the S3 blobstore.

commons-compress has fixed this upstream, but it's not in a
released version yet, so I've included a local copy of the
fixed class until this appears in an upstream release and can be
fixed by depending on a newer version.